requirements.txt
cloudpickle==2.2.0
contourpy==1.0.5
crafter==1.8.0
cycler==0.11.0
fonttools==4.38.0
gym==0.26.2
gym-notices==0.0.8
imageio==2.22.2
imageio-ffmpeg==0.4.7
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.0
numba==0.56.3
numpy==1.23.4
opensimplex==0.4.3
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
seaborn==0.12.1
six==1.16.0
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.4.0
